+++ worlds/pokemon_bw/options/encounter.py
@@ -14,54 +14,43 @@ if typing.TYPE_CHECKING:
 
 class RandomizeWildPokemon(ToggleSet):
     """
-    Randomizes wild pokemon encounters.
-    You can add as many of the following modifiers as you want.
+    Randomizes wild pokemon encounters. You can add as many of the following modifiers as you want.
 
     The following is an example for how options like this can look like (when edited in a yaml):
     ```
     randomize_wild_pokemon:
       ["Randomize", "Prevent rare encounters"]
 
-    ```
-    Here is an alternative way to format it:
-    ```
+    # Or alternatively...
     randomize_wild_pokemon:
     - Randomize
     - Prevent rare encounters
 
     ```
 
-    - **Randomize** - Toggles wild pokemon being randomized. Automatically added if any
-        other modifier is added.
-    - **Ensure all obtainable** - Ensures that every pokemon species is obtainable by
-        either catching or evolving.
-    - **Similar base stats** - Tries to keep every randomized pokemon at a similar base
-        stat total as the replaced encounter.
-    - **Prevent overpowered pokemon** - Tries to prevent pokemon with a base stat total
-        over an adjustable threshold being randomized into wild encounter slots. Other
-        modifiers (except for **Similar base stats**) take priority in case of conflicts.
-    - **Prevent bad early pokemon** - Prevents encountering pokemon with Wonder Guard or
-        fixed HP attacks in areas that are not that far from home. This includes all
-        areas before Castelia City. Might not be fully ensured depending on other options.
-    - **Type themed areas** - Tries to make every pokemon in an area have a certain same
-        type. Might not be fully ensured depending on RNG.
-    - **Area 1 to 1** - Keeps the amount of different encounters and their encounter rate
-        in every area.
-    - **Dungeon 1 to 1** - Keeps the amount of different encounters and their encounter
-        rate across all areas (and if it's part of a dungeon, then across all areas of
-        that dungeon). Supersedes **Area 1 to 1**.
-    - **Global 1 to 1** - Keeps the amount of different encounters and their encounter
-        rate across all areas in the game. Supersedes **Area 1 to 1** and
-        **Dungeon 1 to 1**. Ignored when **Ensure all obtainable** is added.
-    - **Merge phenomena** - Makes rustling grass, rippling water spots, dust clouds,
-        and flying shadows in the same area have only one encounter. Takes priority over
-        **Area 1 to 1** and alike.
-    - **Prevent rare encounters** - Randomizes the encounter slots with the lowest chance
-        in each area to the same pokemon. Takes priority over **Area 1-to-1** and alike.
+    - **Randomize** - Toggles wild pokemon being randomized. Automatically added if any other modifier is added.
+    - **Ensure all obtainable** - Ensures that every pokemon species is obtainable by either catching or evolving.
+    - **Similar base stats** - Tries to keep every randomized pokemon at a similar base stat total as the replaced encounter.
+    - **Prevent overpowered pokemon** - Tries to prevent pokemon with a base stat total over an adjustable threshold
+        being randomized into wild encounter slots. Other modifiers (except for **Similar base stats**) take priority in
+        case of conflicts.
+    - **Prevent bad early pokemon** - Prevents encountering pokemon with Wonder Guard or fixed HP attacks in areas that
+        are not that far from home. This includes all areas before Castelia City. Might not be fully ensured depending
+        on other options.
+    - **Type themed areas** - Tries to make every pokemon in an area have a certain same type. Might not be fully
+        ensured depending on RNG.
+    - **Area 1 to 1** - Keeps the amount of different encounters and their encounter rate in every area.
+    - **Dungeon 1 to 1** - Keeps the amount of different encounters and their encounter rate across all areas (and if
+        it's part of a dungeon, then across all areas of that dungeon). Supersedes **Area 1 to 1**.
+    - **Global 1 to 1** - Keeps the amount of different encounters and their encounter rate across all areas in the
+        game. Supersedes **Area 1 to 1** and **Dungeon 1 to 1**. Ignored when **Ensure all obtainable** is added.
+    - **Merge phenomena** - Makes rustling grass, rippling water spots, dust clouds, and flying shadows in the same
+        area have only one encounter. Takes priority over **Area 1 to 1** and alike.
+    - **Prevent rare encounters** - Randomizes the encounter slots with the lowest chance in each area to the same
+        pokemon. Takes priority over **Area 1-to-1** and alike.
 
-    It is **highly recommended** to include **Prevent rare encounters** or one of the
-    **... 1 to 1** modifiers if you want to randomize wild pokemon, else you might find
-    yourself searching for multiple 1% encounters on every route.
+    It is **highly recommended** to include **Prevent rare encounters** or one of the **... 1 to 1** modifiers if you
+    want to randomize wild pokemon, else you might find yourself searching for multiple 1% encounters on every route.
     """
     # **Ensure all obtainable** -  ... This is automatically checked if **National pokedex** is chosen as the goal.
     # **Prevent bad early pokemon** - ... Without door shuffle, this includes...
@@ -90,29 +79,28 @@ class RandomizeTrainerPokemon(ToggleSet):
     Randomizes trainer pokemon.
     You can add as many of the following modifiers as you want.
 
-    - **Randomize** - Toggles trainer pokemon being randomized. Automatically added if
-        any other modifier is added.
-    - **Similar base stats** - Tries to keep the randomized pokemon at a similar base
-        stat total as the replaced one.
-    - **Prevent overpowered pokemon** - Prevents trainers from having pokemon with a base
-        stats total above an adjustable threshold. Takes priority over most other modifiers.
-    - **Evolve when possible** - Tries to evolve pokemon if they are able to (based on
-        their level). Pokémon that evolve independently of their level are evolved at level 25.
+    - **Randomize** - Toggles trainer pokemon being randomized. Automatically added if any other
+        modifier is added.
+    - **Similar base stats** - Tries to keep the randomized pokemon at a similar base stats
+        total as the replaced one.
+    - **Prevent overpowered pokemon** - Prevents trainers from having pokemon with a base stats
+        total above an adjustable threshold. Takes priority over most other modifiers.
+    - **Evolve when possible** - Tries to evolve pokemon if they are able to (based on their
+        level). Pokémon that evolve independently of their level are evolved at level 25.
     - **Force fully evolved** - Always fully evolves pokemon above a certain (adjustable) level.
     - **Force not underpowered** - Prevents pokemon above a certain (adjustable) level from
         having base stats below a certain (adjustable) total.
     - **Type themed** - All pokemon of a trainer will share at least one randomly chosen type.
-    - **Shuffle gym leader types** - Assigns a (unique) random type to each gym leader
-        and elite 4 member instead of using their vanilla type. Do note that they always
-        have type themed teams.
-    - **Themed gym trainers** - All pokemon of gym trainers will share the type assigned
-        to their gym leader.
-    - **Rivals keep starter** - Makes all Bianca/Cheren fights have one pokemon in
-        common, which will always evolve when possible.
+    - **Shuffle gym leader types** - Assigns a (unique) random type to each gym leader and elite 4
+        member instead of using their vanilla type. Do note that they always have type themed teams.
+    - **Themed gym trainers** - All pokemon of gym trainers will share the type assigned to their
+        gym leader.
+    - **Rivals keep starter** - Makes all Bianca/Cheren fights have one pokemon in common, which
+        will always evolve when possible.
 
-    Combining multiple modifiers that affect all trainer pokémon (i.e. not just gym
-    trainers, ...) might massively increase generation time. Also in that case, most
-    options tend to be less strict in order to reduce generation time.
+    Combining multiple modifiers that affect all trainer pokémon (i.e. not just gym trainers, ...)
+    might massively increase generation time. Also in that case, most options tend to be less
+    strict in order to reduce generation time.
     """
     display_name = "Randomize Trainer Pokemon"
     is_randomize = False
@@ -139,16 +127,19 @@ class RandomizeStarterPokemon(ToggleSet):
     Randomizes the starter pokemon you receive at the start of the game.
     You can add as many of the following modifiers as you want.
 
-    - **Randomize** - Toggles starter pokemon being randomized. Automatically added if
-        any other modifier is added.
+    - **Randomize** - Toggles starter pokemon being randomized.
+        Automatically added if any other modifier is added.
     - **Any base** - Only use unevolved/baby pokemon.
-    - **Base with 2 evolutions** - Only use unevolved/baby pokemon that can evolve twice
-        (or more, if evolutions are randomized). Overrides **Any base**.
-    - **Only official starters** - Only use pokemon that have been a starter in any
-        mainline game. Overrides **Any base** and **Base with 2 evolutions**.
-    - **Type variety** - Every starter will have types that are different from the other two.
-    - **Vanilla types** - Every starter will have one of the vanilla types, i.e. Grass,
-        Fire, and Water.
+    - **Base with 2 evolutions** - Only use unevolved/baby pokemon
+        that can evolve twice (or more, if evolutions are randomized).
+        Overrides  **Any base**.
+    - **Only official starters** - Only use pokemon that have been
+        a starter in any mainline game. Overrides **Any base** and
+        **Base with 2 evolutions**.
+    - **Type variety** - Every starter will have types that are
+        different  from the other two.
+    - **Vanilla types** - Every starter will have one of the vanilla
+        types, i.e. Grass, Fire, and Water.
     """
     display_name = "Randomize Starter Pokemon"
     is_randomize = False
@@ -163,19 +154,21 @@ class RandomizeStarterPokemon(ToggleSet):
 
 class RandomizeStaticPokemon(ToggleSet):
     """
-    Randomizes static encounters you can battle and catch throughout the game,
-    e.g. Volcarona in Relic Castle.
+    Randomizes static encounters you can battle and catch throughout
+    the game, e.g. Volcarona in Relic Castle.
     You can add as many of the following modifiers as you want.
 
-    - **Randomize** - Toggles static pokemon being randomized. Automatically added if any
-        other modifier is added.
-    - **Similar base stats** - Tries to keep the randomized pokemon at a similar base
-        stat total as the replaced one.
+    - **Randomize** - Toggles static pokemon being randomized.
+        Automatically added if any other modifier is added.
+    - **Similar base stats** - Tries to keep the randomized pokemon
+        at a similar base stat total as the replaced one.
     - **Any base** - Only use unevolved/baby pokemon.
-    - **No legendaries** - Exclude legendaries from being placed into static encounters.
-    - **Split statues** - Splits the statues in Desert Resort into 5 different species.
-    - **Split mimics** - Splits the fake pokeballs on routes 6 and 10 into more than just
-        two different pokemon.
+    - **No legendaries** - Exclude legendaries from being placed
+        into static encounters.
+    - **Split statues** - Splits the statues in Desert Resort into
+        5 different species.
+    - **Split mimics** - Splits the fake pokeballs on routes 6 and
+        10 into more than just two different pokemon.
     """
     display_name = "Randomize Static Pokemon"
     is_randomize = False
@@ -189,16 +182,19 @@ class RandomizeStaticPokemon(ToggleSet):
 
 class RandomizeGiftPokemon(ToggleSet):
     """
-    Randomizes gift pokemon that you receive for free, e.g. the Larvesta egg on route 18.
+    Randomizes gift pokemon that you receive for free, e.g. the
+    Larvesta egg on route 18.
     You can add as many of the following modifiers as you want.
 
-    - **Randomize** - Toggles gift pokemon being randomized. Automatically added if any
-        other modifier is added.
-    - **Similar base stats** - Tries to keep the randomized pokemon at a similar base
-        stat total as the replaced one.
-    - **No legendaries** - Exclude legendaries from being placed into gift encounters.
-    - **Split monkeys** - Makes the gift encounter in Dreamyard depend on which starter
-        you picked (like in vanilla), else it will always give you the same species.
+    - **Randomize** - Toggles gift pokemon being randomized.
+        Automatically added if any other modifier is added.
+    - **Similar base stats** - Tries to keep the randomized
+        pokemon at a similar base stat total as the replaced one.
+    - **No legendaries** - Exclude legendaries from being placed
+        into gift encounters.
+    - **Split monkeys** - Makes the gift encounter in Dreamyard
+        depend on which starter you picked (like in vanilla),
+        else it will always give you the same species.
     """
     display_name = "Randomize Gift Pokemon"
     is_randomize = False
@@ -210,16 +206,16 @@ class RandomizeGiftPokemon(ToggleSet):
 
 class RandomizeTradePokemon(ToggleSet):
     """
-    Randomizes trade offers from NPCs. Any **Randomize ...** is required for the
-    other modifiers.
+    Randomizes trade offers from NPCs. Any **Randomize ...**
+    is required for the other modifiers.
     You can add as many of the following modifiers as you want.
 
     - **Randomize offer** - Toggles offered pokemon being randomized.
     - **Randomize request** - Toggles requested pokemon being randomized.
-    - **Similar base stats** - Tries to keep the randomized pokemon at a similar base
-        stat total as the replaced one.
-    - **Coupled base stats** - Tries to make offered and requested pokemon have similar
-        base stats.
+    - **Similar base stats** - Tries to keep the randomized pokemon at a
+        similar base stats total as the replaced one.
+    - **Coupled base stats** - Tries to make offered and requested
+        pokemon have similar base stats.
     - **No legendaries** - Exclude legendaries from being placed into trades.
     """
     display_name = "Randomize Trade Pokemon"
@@ -235,16 +231,19 @@ class RandomizeLegendaryPokemon(ToggleSet):
     Randomizes legendary and mythical encounters.
     You can add as many of the following modifiers as you want.
 
-    - **Randomize** - Toggles legendary pokemon being randomized. Automatically added if
-        any other modifier is added.
-    - **Keep legendary** - Randomized pokemon will all still be legendaries or mythicals.
-    - **No legendaries** - Exclude legendaries from being placed into these encounters.
-    - **Similar base stats** - Tries to keep the randomized pokemon at a similar base
-        stat total as the replaced one. Overrides **Keep legendary**.
+    - **Randomize** - Toggles legendary pokemon being randomized.
+        Automatically added if any other modifier is added.
+    - **Keep legendary** - Randomized pokemon will all still be
+        legendaries or mythicals.
+    - **No legendaries** - Exclude legendaries from being placed
+        into these encounters.
+    - **Similar base stats** - Tries to keep the randomized
+        pokemon at a similar base stat total as the replaced one.
+        Overrides **Keep legendary**.
     - **Same type** - Tries to keep at least one type of every encounter.
 
-    Including **Keep legendary** AND **No legendaries** will instead only put pseudo
-    legendaries into these encounters.
+    Including **Keep legendary** AND **No legendaries** will instead
+    only put pseudo legendaries into these encounters.
     """
     display_name = "Randomize Legendary Pokemon"
     is_randomize = False
@@ -257,23 +256,25 @@ class RandomizeLegendaryPokemon(ToggleSet):
 
 class PokemonRandomizationAdjustments(ExtendedOptionCounter):
     """
-    Adjust various parameters in various pokemon randomization options
-    (with individual ranges).
-    Every parameter can be specified as unweighted/weighted lists, "random",
-    and "random-range-x-y" like usual range options.
+    Adjust various parameters in various pokemon randomization
+    options (with individual ranges).
+    Every parameter can be specified as unweighted/weighted lists,
+    "random", and "random-range-x-y" like usual range options.
 
-    - **Stats leniency** (0-1530) - The starting maximum difference between base stat
-        totals of vanilla and randomized species (for options with **Similar base stats**
-        activated).
-    - **Rare encounters threshold** (1-100) - If **Prevent rare encounters** is included,
-        this will be the minimum encounter chance (in percent) for each species.
-    - **Overpowered threshold** (200, 1530) - The maximum base stat total (for options
-        with **Prevent overpowered pokemon** activated).
-    - **Underpowered threshold** (200, 1530) - The minimum base stat total (for options
-        with **Force not underpowered** activated).
+    - **Stats leniency** (0-1530) - The starting maximum difference
+        between base stat totals of vanilla and randomized species
+        (for options with **Similar base stats** activated).
+    - **Rare encounters threshold** (1-100) - If **Prevent rare encounters**
+        is included, this will be the minimum encounter chance (in percent)
+        for each species.
+    - **Overpowered threshold** (200, 1530) - The maximum base stat total
+        (for options with **Prevent overpowered pokemon** activated).
+    - **Underpowered threshold** (200, 1530) - The minimum base stat total
+        (for options with **Force not underpowered** activated).
     - **Force threshold** (1, 100) - The minimum level at which trainer
-        pokemon are forced to be fully evolved (if **Force fully evolved** is included)
-        or not be underpowered (if **Force not underpowered** is included).
+        pokemon are forced to be fully evolved (if **Force fully evolved**
+        is included) or not be underpowered (if **Force not underpowered**
+        is included).
     """
     display_name = "Pokemon Randomization Adjustments"
     fill_defaults = True
@@ -331,9 +332,9 @@ class EncounterPlando(Option[list[PlandoEncounter]]):
       species: Name(s) of species, random if multiple
 
     ```
-    Encounter Plando requires the corresponding host setting to be enabled, else it will
-    be ignored for all players. Be aware that this can lead to generation failures when
-    combined with other restrictive options.
+    Encounter Plando requires the corresponding host setting to be enabled,
+    else it will be ignored for all players. Be aware that this can lead to
+    generation failures when combined with other restrictive options.
     Refer to the Encounter Plando guide of this game for further information.
     """
     display_name = "Encounter Plando"
@@ -481,8 +482,9 @@ class EncounterPlando(Option[list[PlandoEncounter]]):
 class WildRandomizationBlacklist(OptionSet):
     """
     Excludes a list of pokemon from being used in wild randomization.
-    Be aware that certain pokemon still have to be encountered somewhere, especially with
-    **Ensure all obtainable** enabled. Also, a big list can lead to generation failures.
+    Be aware that certain pokemon still have to be encountered somewhere,
+    especially with **Ensure all obtainable** enabled. Also, a big list
+    can lead to generation failures.
     """
     display_name = "Wild Randomization Blacklist"
     valid_keys = list(species.by_name)